requirements.txt
fastapi
uvicorn
sqlalchemy
aiomysql
passlib[bcrypt]
python-jose[cryptography]
celery
redis
pydantic